Move codes under /src dir
